# openapi.cloudflare.yaml
openapi: 3.0.3
info:
  title: ExitRoute API
  description: A self-hosted informational API. It never authenticates to, operates, or stores credentials
    for a consumer account.
  version: 0.1.0
paths:
...
  /v1/observations:
    post:
      tags:
      - observations
      summary: Create Observation
      operationId: createObservation
      security:
      - ApiKeyAuth: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationReceipt'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /admin/v1/observations:
    get:
      tags:
      - admin
      summary: Admin List Observations
      operationId: adminListObservations
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: moderation_state
        in: query
        required: false
        schema:
          enum:
          - queued
          - accepted
          - rejected
          type: string
          default: queued
          title: Moderation State
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationAdminPage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  schemas:
...
    ObservationAdminPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ObservationAdminView'
          type: array
          title: Data
        next_cursor:
          title: Next Cursor
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - data
      - next_cursor
      title: ObservationAdminPage
    ObservationAdminView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        client_id:
          type: string
          format: uuid
          title: Client Id
        payload:
          additionalProperties: true
          type: object
          title: Payload
        moderation_state:
          type: string
          enum:
          - queued
          - accepted
          - rejected
          title: Moderation State
        received_at:
          type: string
          format: date-time
          title: Received At
        moderated_at:
          title: Moderated At
          type: string
          format: date-time
          nullable: true
        moderated_by:
          title: Moderated By
          type: string
          nullable: true
        decision_reason:
          title: Decision Reason
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - id
      - client_id
      - payload
      - moderation_state
      - received_at
      - moderated_at
      - moderated_by
      - decision_reason
      title: ObservationAdminView
    ObservationCreate:
      properties:
        service_slug:
          type: string
          maxLength: 80
          minLength: 2
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          title: Service Slug
        observed_revision:
          title: Observed Revision
          type: integer
          minimum: 1.0
          nullable: true
        outcome:
          $ref: '#/components/schemas/Outcome'
          default: cancel_subscription
        variant:
          $ref: '#/components/schemas/Variant'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        change_types:
          items:
            type: string
            enum:
            - entry_url_changed
            - choice_label_changed
            - step_added
            - step_removed
            - loop_added
            - retention_offer_added
            - offline_handoff_added
            - cancellation_terminal_missing
            - variant_mismatch
            - other
          type: array
          maxItems: 10
          minItems: 1
          title: Change Types
        note:
          title: Note
          type: string
          maxLength: 500
          nullable: true
      additionalProperties: false
      type: object
      required:
      - service_slug
      - occurred_at
      - change_types
      title: ObservationCreate
...
    ObservationReceipt:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        moderation_state:
          type: string
          enum:
          - queued
          - accepted
          - rejected
          title: Moderation State
          default: queued
        duplicate:
          type: boolean
          title: Duplicate
        received_at:
          type: string
          format: date-time
          title: Received At
      additionalProperties: false
      type: object
      required:
      - id
      - duplicate
      - received_at
      title: ObservationReceipt
    Outcome:
      type: string
      enum:
      - cancel_subscription
      title: Outcome
    Platform:
      type: string
      enum:
      - web
      - ios
      - android
      - phone
      - mail
      - in_person
      title: Platform
...
    Variant:
      properties:
        region:
          type: string
          pattern: ^[A-Z]{2}$
          title: Region
          default: US
        platform:
          $ref: '#/components/schemas/Platform'
          default: web
      additionalProperties: false
      type: object
      title: Variant
...
    Problem:
      additionalProperties: false
      properties:
        type:
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          title: Detail
          type: string
        code:
          title: Code
          type: string
        instance:
          default: null
          title: Instance
          type: string
          nullable: true
        request_id:
          default: null
          title: Request Id
          type: string
          nullable: true
        errors:
          default: null
          title: Errors
          items:
            additionalProperties: true
            type: object
          type: array
          nullable: true
      required:
      - type
      - title
      - status
      - detail
      - code
      title: Problem
      type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: A scoped key created by an ExitRoute administrator.
      in: header
      name: X-API-Key
    BootstrapAdminAuth:
      type: http
      description: Bootstrap-only bearer token. Prefer a scoped admin API key after setup.
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the required scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: The request conflicts with resource state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: The request violates a semantic or safety rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The API-client quota was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Seconds until a new quota window.
          schema:
            type: integer
            minimum: 1
    InternalError:
      description: The server could not complete the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...